requirements.txt
pip~=22.3.1
parso~=0.8.1
requests~=2.27.1
urllib3~=1.26.9
python-dotenv~=0.9.1